+++ packages/forest_admin_agent/spec/lib/forest_admin_agent/routes/resources/update_field_spec.rb
@@ -10,6 +10,7 @@ module ForestAdminAgent
 
       describe UpdateField do
         include_context 'with caller'
+        include_context 'with readable related collections'
         subject(:update_field) { described_class.new }
 
         let(:args) do