+++ packages/forest_admin_datasource_active_record/lib/forest_admin_datasource_active_record/parser/column.rb
@@ -17,6 +17,9 @@ module ForestAdminDatasourceActiveRecord
         string: 'String',
         text: 'String',
         citext: 'String',
+        inet: 'String',
+        cidr: 'String',
+        macaddr: 'String',
         time: 'Time',
         uuid: 'Uuid',
         binary: 'Binary'